refactor: Game/Room으로 Controller 분리
